Split stuff from IsoFileCore
