+++ vssource/formats/dvd/__init__.py
@@ -1,2 +1,3 @@
 from . import parsedvd  # noqa: F401,F403
 from .IsoFile import *  # noqa: F401,F403
+from .title import *  # noqa: F401,F403
